色々修正

+++ NjTest/NinjaSprit/System/File.cpp
@@ -16,7 +16,7 @@ tag_(tag)
 		auto fsize = endIdx - bgnIdx;
 		FileRead_seek(handle, 0, SEEK_SET);
 		data_.resize(fsize);
-		FileRead_read(data_.data(), fsize, handle);
+		FileRead_read(data_.data(), static_cast<int>(fsize), handle);
 		FileRead_close(handle);
 		cursor_ = data_.begin();
 	}
